feat : 구조 개선을 위해 관련 코드 배치 수정
