+++ src/utils/interval-util.ts
@@ -1,0 +1,7 @@
+export function startDelayedInterval(callback, intervalTime, maxDelay) {
+  setTimeout(function () {
+    setInterval(function () {
+      callback();
+    }, intervalTime);
+  }, Math.random() * maxDelay);
+}
